+++ driver/sql/postgres/pgjournal/schema.sql
@@ -4,6 +4,7 @@ CREATE TABLE
     IF NOT EXISTS persistencekit.journal (
         id BIGSERIAL NOT NULL,
         name TEXT NOT NULL,
+        encoded_begin BIGINT NOT NULL DEFAULT -1::BIGINT << 63, -- see `bigint` package
         PRIMARY KEY (id),
         UNIQUE (name)
     );